fix: Restrict dry-run feature to Zeta CLI and block REST API access
- Remove --dry-run option from AbstractCommandArgs to prevent usage in Flink/Spark engines
- Add --dry-run option to ClientCommandArgs for Zeta CLI exclusive support
- Implement validation in JobInfoService to reject dry-run requests via REST API
- Update technical documentation and tests to reflect the feature boundary
